Added post.go and modified the database.
New file post.go currently only contains the `type Post struct`. The
database.go file now creates a table for posts (if it doesn't exist
already) when initiating the database. Instead of creating a LengthX
func for every table in the database, now we only have one called
Length that takes the name of the table as a parameter and returns the
number of things in whatever table you're looking in.